Fix const issues related to Observable<T>
So once upon a time, I clearly came across some issues
related to const-correctness, logical const, and the
difference between some older version of GCC and
MSVC. This resulted in the definitely correct solution
of using a #ifdef to handle the different compilers
differently.

But now I know better. What I really want here is to
express logical constness; the underlying T value should
behave according to proper const rules, but we don't
care about any of the stats that get gathered so those
can just all be marked mutable, and the rest of the
class can be implemented with const where you'd
expect in any sane type.

This doesn't properly handle rvalues and move
semantics, which is fine for now because all the values
are numbers; but it'd be nice to solve that in the
future. This at least gets things building again with
modern compilers.