Add glut::Window to manage callbacks
One of the main frustrations I have with the glut library
is that it makes it pretty difficult to handle callbacks
without relying on some sort of global state. This
Window class encapsulates as much of that as possible
to store and invoke std::function objects as callbacks
within the globally registered glut callbacks; so from
the rest of the program, we can move to handling
callbacks without needing to rely on global state.