Reorganize source tree
The long term plan here is to have two separate parts to this library -
sort algorithms which could be used in a real world project, and the
visualizer. This means a couple things
* The visualizer needs to get rewritten to not require hooks in the sort
algorithm (which would allow viewing std::sort or other algorithms)
* The sort algorithms need to be rewritten to have an interface more like
std::sort
* The algorithms need to be moved out of where the main visualizer is
so that they could be included alone

This also means it'll be possible to unit test the algorithms, which
is sorely needed (quicksort currently doesn't work for sets with
duplicates)

This starts the whole process off by reorganizing the source tree to nest
all the existing code down a level so the sort algorithms can be
moved/rebuilt in a different directory.