Add new bucket sort algorithm
This replaces LSB radix sort. Since it's defined in terms of a bucket
traits type, it can be specialized to work with strings as well.

This could be further improved by allocating all buckets in a single
list, but then we need to keep track of which items have been constructed
or not. For now, a separate vector per bucket will do, we're still only
allocating at the beginning of the sort, not each time through.