Add back CLI support for picking visualization
This uses a bit different approach to rendering than previously - instead
of one function knowing how to render the whole list, this uses a
strategy approach. However, since the strategies don't use runtime
polymorphism, it's still not a super clean solution, so this may not be
worth it in the long run.

Either way, the draw function now gets defined by the command line
arguments again, and main() doesn't know the details of the specific
options.